# perf(npm): don't try to cache npm packages we've already cached