@dsherret 2025-04-17
Maybe this solves https://github.com/denoland/deno/issues/28933 ?