@nathanwhit 2025-04-17
> Maybe this solves #28933 ? (Or at least partly)

Yeah it should help reduce the spam, but not a proper fix